requirements.txt
PyQt6
pywin32
pydantic-settings
pytesseract
pillow
python-telegram-bot